Do not update Agreements through supplier_framework route
All updates to FrameworkAgreements should now be done through the FrameworkAgreement routes, so while the SupplierFramework serialization still contains FrameworkAgreement information it is no longer possible to update agreement info this way.

The route has been renamed to reflect the fact that the only thing that can be updated via this route is whether a supplier is on a framework or not.