Add initial api spec in OpenAPI v3 draft format
Adds an API description file in the OpenAPI v3 draft format.

The specification is generated from the API requests generated
by the API test suite, and describes the list of tested endpoints,
parameters and example requests and responses.

I've checked that the endpoints match the list generated by `list_routes`
management command and that the file passes OpenAPI v3 validation.

It's missing useful summaries and descriptions for the endpoints,
responses and examples. It also contains most of the requests sent by
the tests, so a lot of the examples are repeating or aren't useful.